Lots of changes made just to make the code fit all our diagrams for iteration one.
A lot of review needs to be done.  

Most notable change: GameWolds.map is not public. Obviously this shouldn't be the case.  It was changed to make the code compile.  As it is right now Monirator is a singleton for validating requests.  Monirator will need access to tilemap among other things to validate data.  It should be that getting map should be public but seting map should be private.